Refactor: Rename StudyLog package to Mission package
